frontend/requirements.txt
litellm
google-adk